push: Move FlattenedJson to its own module
